refactor(es): Extract rotation types to config_rotation.go
Move RotationConfig, PeriodicRotation, ManualRolloverRotation,
AutoRolloverRotation, DataStreamRotation structs, HasRotation method,
validateRotationConfig, and legacyRotationFlagsList constant to a
dedicated file for better organization.
